Allowing the language to be passed the filepath to the language file directly - allowing the behaviour to be the same as using the tags such as 'en' etc but from different locations.
Revert "Allowing the language to be passed the filepath to the language file directly - allowing the behaviour to be the same as using the tags such as 'en' etc but from different locations."

This reverts commit 158af014d724be349ea55b942d07b5e987309abe.

Allowing the language arg to take the usual tag such as 'en' but also to take an absolute path to the language file whilst behaving in the same way.

By allowing the argument to take in an absolute path, it gives the module flexibility to integrate into projects with rigid data structures, where the languages will be placed in alternative locations.